Convert UsersController to use module prepend
